+++ src/Streamly/Internal/Data/Stream/IsStream/Generate.hs
@@ -75,7 +75,6 @@ module Streamly.Internal.Data.Stream.IsStream.Generate
     , fromListM
     , K.fromFoldable
     , fromFoldableM
-    , fromPrimIORef
     , fromCallback
 
     -- * Deprecated
@@ -102,8 +101,8 @@ import Streamly.Internal.Data.Stream.StreamK (IsStream((|:), consM))
 import Streamly.Internal.Data.Stream.Serial (SerialT, WSerialT)
 import Streamly.Internal.Data.Stream.Zip (ZipSerialM)
 import Streamly.Internal.Data.Time.Units (AbsTime , RelTime64, addToAbsTime64)
-import Streamly.Internal.Data.IORef.Prim (Prim, IORef)
 
+import qualified Streamly.Internal.Data.Stream.Parallel as Par
 import qualified Streamly.Internal.Data.Stream.Prelude as P
 import qualified Streamly.Internal.Data.Stream.Serial as Serial
 import qualified Streamly.Internal.Data.Stream.StreamD.Generate as D
@@ -540,14 +539,8 @@ fromHandle h = go
             str <- liftIO $ IO.hGetLine h
             yld str go
 
--- | Construct a stream by reading a 'Prim' 'IORef' repeatedly.
+-- XXX This should perhaps be moved to Parallel
 --
--- /Internal/
---
-{-# INLINE fromPrimIORef #-}
-fromPrimIORef :: (IsStream t, MonadIO m, Prim a) => IORef a -> t m a
-fromPrimIORef = fromStreamD . D.fromPrimIORef
-
 -- | Takes a callback setter function and provides it with a callback.  The
 -- callback when invoked adds a value at the tail of the stream. Returns a
 -- stream of values generated by the callback.
@@ -557,6 +550,6 @@ fromPrimIORef = fromStreamD . D.fromPrimIORef
 {-# INLINE fromCallback #-}
 fromCallback :: MonadAsync m => ((a -> m ()) -> m ()) -> SerialT m a
 fromCallback setCallback = concatM $ do
-    (callback, stream) <- D.newCallbackStream
+    (callback, stream) <- Par.newCallbackStream
     setCallback callback
     return stream
