+++ src/Streamly/Internal/Data/Unfold.hs
@@ -164,8 +164,10 @@ import qualified Prelude
 import qualified Control.Monad.Catch as MC
 import qualified Data.Tuple as Tuple
 import qualified Streamly.Internal.Data.Fold.Types as FL
-import qualified Streamly.Internal.Data.Stream.StreamK as K
+import qualified Streamly.Internal.Data.Stream.StreamK as K (uncons)
+import qualified Streamly.Internal.Data.Stream.StreamK.Type as K
 import qualified Streamly.Internal.Data.Stream.StreamD as D
+    (unfold, newFinalizedIORef, runIORefFinalizer, withIORefFinalizer)
 
 import Streamly.Internal.Data.SVar
 import Prelude
